Update patch set 1
Patch Set 1:

(1 comment)

It looks like a lot of this is default ONAP stuff. Would it be possible for you to first do a commit of just the totally unmodified ONAP stuff that you presumably downloaded, then a separate commit with just the additions and changes that you made?

Patch-set: 1
Label: Code-Review=0