Update patch set 2
Patch Set 2: Code-Review-1

Since you've confirmed that this is mostly ONAP code I'm definitely -1 on this commit as-is. Please submit a commit that is entirely unmodified ONAP code first (and as Bob said, make sure the commit message clearly states that it is unmodified ONAP code). Then after that commit is merged, we can review a separate commit that contains only your changes to the ONAP code.

In the commit message for the ONAP code, include an exact reference for where it came from (i.e. the git URL and specific commit hash, or a tar file URL and sha256sum, or whatever you used as your starting point before you started making changes)

I want the git log and git blame for this repository to clearly show the provenance of the code.

Patch-set: 2
Label: Code-Review=-1