Update patch set 1
Patch Set 1: Code-Review-1

Are these files going to stay in this location?
It is more normal to have the pom.xml and src/ directory in the root of the repo.
Otherwise you have to tell the build system into finding the pom.xml.  I would
move the contents of the directory up to the root level (e.g. get rid of epsdk-app-os directory level)

Patch-set: 1