Update patch set 1
Patch Set 1:

> Are these files going to stay in this location?
 > It is more normal to have the pom.xml and src/ directory in the
 > root of the repo.
 > Otherwise you have to tell the build system into finding the
 > pom.xml.  I would
 > move the contents of the directory up to the root level (e.g. get
 > rid of epsdk-app-os directory level)

I want the final structure of the portal_user_interface/portal-onapsdk project to be the following:

ARCPORTAL-SDK-APP (subproject)
ONAP-SDK-APP (subproject)
pom.xml (parent pom.xml)

The ONAP-SDK-APP will be based on the ONAP portal epsdk-app-os subproject. I want limited changes to be applied to that subproject in order for potential ONAP portal SDK updates to be easily applied in the future. All the ARC portal functionality will be located in ARCPORTAL-SDK-APP subproject which will be an overlay for the ONAP-PORTAL-SDK subproject.

So, I copied the epsdk-app-os subproject from ONAP repo and placed in under the portal-onapsdk repo. The next step will be to rename this subproject to 'ONAP-SDK-APP', apply limited changes in order to meet our requirements, create the 'ARCPORTAL-SDK-APP' subporject and the paretn pom.xml

Also, by using this project structure, we are aligned with the ONAP portal SDK structure.

Patch-set: 1
Reviewer: Gerrit User 194 <194@4cada01b-0314-4fe0-84a1-f9ddc89863f9>
Label: Verified=0